requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pdfplumber
pytesseract
pillow
google-generativeai
firebase-admin